Change 'Newest First' sort order to use 'weight' field and not releaseDate
BC releaseDate field has "issues". It cannot be updated via product import, so has to be manually set. Plus, a bug in the import resets this field to 1 Jan 2000 each time you import, so if you do set the releaseDate field manually in BC Admin, it's only going to get overwritten the next time you do an import. Usual BC rubbish...Hopefully the 'weight' field will be a good subtitute as it can be set via the import and gives us a measure of control over the sort order if we give newer items a larger 'weight'